requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langchain-community
langchain-core
langchain-text-splitters
sentence-transformers
chromadb
pydantic
python-dotenv
jinja2
graphviz
pygame
requests
pillow
qdrant-client
wikipedia
pytest
streamlit
python-multipart
torch
transformers